pyproject.toml
[project]
name = "CoverUp"
description = "LLM-powered test coverage improver"
readme = "README.md"
version = "0.1.2"
authors = [
    { name="Juan Altmayer Pizzorno", email="juan@altmayer.com" },
    { name="Emery Berger", email="emery.berger@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X"
]
requires-python = ">=3.10"
dependencies = [
    "asyncio",
    "openai",
    "tiktoken",
    "aiolimiter",
    "tqdm",
    "slipcover>=1.0.8",
    "pytest-forked",
    "litellm>=1.33.1"
]

[project.scripts]
coverup = "coverup.coverup:main"

[project.urls]
"Homepage" = "https://github.com/plasma-umass/CoverUp"
